# verificar-front/src/generated/openapi/model/documentoResponse.ts
/**
 * ine-verificator
 * API para verificar documentos oficiales como INE, CSF y documentos bancarios.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: soporte@ejemplo.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TipoDocumento } from './tipoDocumento';
import { ArchivoBase64 } from './archivoBase64';
import { EstadoDocumento } from './estadoDocumento';


export interface DocumentoResponse { 
    /**
     * ID del documento
     */
    documentoId?: string;
    tipo_documento?: TipoDocumento;
    estado?: EstadoDocumento;
    /**
     * Documento en formato de bytes codificado en Base64
     */
    archivo?: ArchivoBase64;
}
export namespace DocumentoResponse {
}